Start separating into multiple ShurikenData
