+++ CMakeLists.txt
@@ -45,6 +45,7 @@ add_library(puzzles_lib OBJECT
         src/cpic/solver/brute_force_board_solver.cpp
         src/cpic/solver/heuristic_board_solver.cpp
         src/cpic/view/board_logger.cpp
+        src/shurikens/data/trivial.cpp
         src/shurikens/logger.cpp
         src/shurikens/solver/breadth_search_solver.cpp
         src/sudoku/data/easy.cpp
